# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api documentation",
        "description": "Ticketing system api documentation",
        "contact": {
            "email": "davit.chechelashvili@rda.gov.ge"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/priority/list": {
            "get": {
                "tags": [
                    "პრიორიტეტების API"
                ],
                "summary": "პრიორიტეტების წამოღების მარსუტი",
                "operationId": "73d87df455e1af65abdcf97191dee7cf",
                "responses": {
                    "200": {
                        "description": "პრიორიტეტები წამოღებულია"
                    },
                    "422": {
                        "description": "პრიორიტეტები არ არსებობს"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/priority/delete/{id}": {
            "delete": {
                "tags": [
                    "პრიორიტეტების API"
                ],
                "summary": "პრიორიტეტების წაშლის მარსუტი",
                "operationId": "b3fcf68864c1f776d98f60a008686e0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "პრიორიტეტის აიდი",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "პრიორიტეტეი წაიშალა"
                    },
                    "422": {
                        "description": "პრიორიტეტი ვერ წაიშალა"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/priority/add": {
            "post": {
                "tags": [
                    "პრიორიტეტების API"
                ],
                "summary": "პრიორიტეტის დამატების მარსუტი",
                "description": "პრიორიტეტის დამატების მეთოდი",
                "operationId": "79fcc10d1fa8cb0a887d041e4d87220d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "პრიორიტეტი დაემატა"
                    },
                    "422": {
                        "description": "პრიორიტეტი ვერ დაემატა"
                    }
                }
            }
        },
        "/api/status/list": {
            "get": {
                "tags": [
                    "სტატუსების API"
                ],
                "summary": "პრიორიტეტების წამოღების მარსუტი",
                "operationId": "33a8510aae252e779cc5d1a1544f5fa0",
                "responses": {
                    "200": {
                        "description": "პრიორიტეტები წამოღებულია"
                    },
                    "422": {
                        "description": "პრიორიტეტები არ არსებობს"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status/delete/{id}": {
            "delete": {
                "tags": [
                    "სტატუსების API"
                ],
                "summary": "სტატუსის წაშლის მარსუტი",
                "operationId": "9dfef41f8e63eda737719bd7128003c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "სტატუსის აიდი",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "სტატუსი წაიშალა"
                    },
                    "422": {
                        "description": "სტატუსი ვერ წაიშალა"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/add": {
            "post": {
                "tags": [
                    "თასქების API"
                ],
                "summary": "თასქის დამატების მარსუტი",
                "description": "თასქის დამატების მეთოდი",
                "operationId": "acffc0c73df6121a2a9d7e63b6a784fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "priority": {
                                        "type": "number",
                                        "format": "number"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "თასქი დაემატა"
                    },
                    "422": {
                        "description": "თასქი ვერ დაემატა"
                    }
                }
            }
        },
        "/api/user/insert": {
            "post": {
                "tags": [
                    "მომხმარებლების API"
                ],
                "summary": "მომხმარებლების ატვირთვის მარშუტი",
                "description": "მომხმარებელთა ატვირთვის მეთოდი",
                "operationId": "e7e568e34b3a248082a6fabf00f2b95b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "position",
                                    "role",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "john doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@simple.com"
                                    },
                                    "position": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "role": {
                                        "type": "number",
                                        "format": "number"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "****"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "მომხმარებლები აიტვირთა"
                    },
                    "422": {
                        "description": "მომხმარებლები ვერ აიტვირთა"
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth",
                        "1": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "ავტორიზაციის API"
                ],
                "summary": "ავტორიზაციის მარშუტი",
                "description": "ავტორიზაციის მეთოდი",
                "operationId": "d89ebc968b603de212f5cb1c5eb34b16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ავტორიზაცია წარმატებით დასრულდა"
                    },
                    "422": {
                        "description": "ავტორიზაცია ვერ განხორციელდა"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "სისტემიდან გამოსვლის API"
                ],
                "summary": "სისტემიდან გამოსვლის მარშუტი",
                "description": "სისტემიდან გამოსვლის მეთოდი",
                "operationId": "269eeaa1f95da774349bbc8f2cbc3c64",
                "responses": {
                    "200": {
                        "description": "Logged out!"
                    },
                    "422": {
                        "description": "Error!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}